% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_lab_data.R
\name{add_solution_id}
\alias{add_solution_id}
\title{Add solution id}
\usage{
add_solution_id(samples)
}
\arguments{
\item{samples}{data.frame with samples in wide format, i.e. parameters are
columns and each row is one samole}
}
\value{
add "solution_id"
}
\description{
Add solution id
}
