% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{convert_simulations_to_list}
\alias{convert_simulations_to_list}
\title{Convert Simulation Results to list}
\usage{
convert_simulations_to_list(simulations)
}
\arguments{
\item{simulations}{tibble with simulation results as rewtrieved by
\code{read_simulations}}
}
\value{
list with simulation results in structure for export to EXCEL
}
\description{
Convert Simulation Results to list
}
