% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_lab_data.R
\name{convert_to_numeric_lab_values}
\alias{convert_to_numeric_lab_values}
\title{Convert to numeric Laboratory Values}
\usage{
convert_to_numeric_lab_values(
  samples_tidy,
  country = "en",
  detLimFactorBelow = 0.5,
  ...
)
}
\arguments{
\item{samples_tidy}{data.frame with samples in tidy format, i.e. parameters are
columns and each row is one value}

\item{country}{"en" if value is given in English format (decimal point ".",
thousands separator ",") or "de" if value is given in German format (decimal
point ",", thousands separator ".").}

\item{detLimFactorBelow}{actor by which detection limit is multiplied in order
to get a valid value when the value was below the detection limit. Default value: 0.5}

\item{...}{additional arguments passed to \link[kwb.base]{hsLabValToVal}}
}
\value{
data.frame with additional columns "outOfLimit" (with "<" or ">" sign)
and "numericValue"
}
\description{
Convert to numeric Laboratory Values
}
