% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_file}
\alias{package_file}
\title{Path to File Stored in Package}
\usage{
package_file(...)
}
\arguments{
\item{\dots}{segments of path to file, in the simplest form just a file name}
}
\value{
full path to file within "extdata" folder of the installed package
}
\description{
Path to File Stored in Package
}
