% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_solutions_input.R
\name{prepare_solutions_input}
\alias{prepare_solutions_input}
\title{Prepare PHREEQC Solutions Input File}
\usage{
prepare_solutions_input(samples_tidy, title = "", min_spaces = 3)
}
\arguments{
\item{samples_tidy}{data.frame with samples in tidy format, i.e. parameters are
columns and each row is one value}

\item{title}{user defined title (default: "")}

\item{min_spaces}{minimum spaces after parameter name based on longest parameter
name (default: 3 spaces)}
}
\value{
input text to be used as PHREEQC input
}
\description{
Prepare PHREEQC Solutions Input File
}
\examples{
file_samples <- system.file("extdata/phreeqc-input.csv",
package = "geosalz.phreeqc")
samples <- read.csv2(file_samples)
samples_tidy <- geosalz.phreeqc::tidy_samples(samples) \%>\%
geosalz.phreeqc::convert_to_numeric_lab_values()
prepare_solutions_input(samples_tidy)

}
