% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output_hauke.R
\name{read_output_file}
\alias{read_output_file}
\title{Read phreeqc output text file into a nested list structure}
\usage{
read_output_file(file)
}
\arguments{
\item{file}{full path to phreeqc output file. Template:
\code{system.file("extdata/phreeqc_output.txt", package =
  "geosalz.phreeqc")}}
}
\description{
Read phreeqc output text file into a nested list structure
}
