% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_redox_couples}
\alias{read_redox_couples}
\title{Helper Function: Read 'Redox Couples' Block}
\usage{
read_redox_couples(txt)
}
\arguments{
\item{txt}{vector of strings with redox_coupkes block}
}
\value{
tibble with redox couples block
}
\description{
Helper Function: Read 'Redox Couples' Block
}
\keyword{internal}
