% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_saturation_indices}
\alias{read_saturation_indices}
\title{Helper Function: Read 'Saturation Indices' Block}
\usage{
read_saturation_indices(txt)
}
\arguments{
\item{txt}{vector of strings with saturation_indices block}
}
\value{
tibble with saturation indices block
}
\description{
Helper Function: Read 'Saturation Indices' Block
}
\keyword{internal}
