% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_simulations}
\alias{read_simulations}
\title{Read Simulations}
\usage{
read_simulations(phreeqc_output)
}
\arguments{
\item{phreeqc_output}{vector with lines of PHREEQC output file}
}
\value{
returns list with elements "end_of_simulations_seconds",input and
output
}
\description{
Read Simulations
}
