% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_lab_data.R
\name{remove_missing_samples}
\alias{remove_missing_samples}
\title{Remove Missing Parameters}
\usage{
remove_missing_samples(
  samples_tidy,
  pattern_missing = "n\\\\.\\\\s?a\\\\.|-",
  col_value = "value"
)
}
\arguments{
\item{samples_tidy}{data.frame with samples in tidy format, i.e. parameters are
columns and each row is one value}

\item{pattern_missing}{remove samples marked "n. a." or "- "}

\item{col_value}{value column (default: "value")}
}
\value{
data.frame without missing samples
}
\description{
Remove Missing Parameters
}
