% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_lab_data.R
\name{tidy_samples}
\alias{tidy_samples}
\title{Tidy Samples}
\usage{
tidy_samples(samples)
}
\arguments{
\item{samples}{data.frame with samples in wide format, i.e. parameters are
columns and each row is one sample}
}
\value{
data.frame with samples in tidy format, i.e. parameters are
columns and each row is one value
}
\description{
Tidy Samples
}
